+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Feature: Added a graph card for the purpose of graphing stock records at various intervals. ([#11](https://github.com/josh-W42/stonks-viewer/pull/11))
 - Feature: Added three new cards, ErrorCard, PlaceHolderCard, and StockInfoCard. Add started the quote page for individual stocks. ([#9](https://github.com/josh-W42/stonks-viewer/pull/9))
 - Fix: Resolved a werid senario invloving navigating when a stock has been selected ([#8](https://github.com/josh-W42/stonks-viewer/pull/8))
 - Feature: Created a simple stock search component. ([#7](https://github.com/josh-W42/stonks-viewer/pull/7))